+++ src/frame/frame.h
@@ -1,5 +1,5 @@
 namespace Frame {
   void setTimestamp(unsigned long time);
 
-  unsigned long getFrameTime();
+  double getFrameTime();
 }
